aoe-sancheck: Refine interface probing
There are two issues in the ethlist function in aoe-sancheck.c:

First it skips all network interfaces that do not start with "eth". This
is a problem these days where people use different naming schemes like the
so-called predictable interface names as mandated by udev/systemd.

Second it only probes for interfaces with an interface number of up
to 15. At least in current Linux, that number can take huge values and
those interfaces will be ignored completely.

About the first: Probe all interfaces (but lo) that support the ARP
protocol, assuming that is a fair indicator for ethernet support.

About the second, use the if_nameindex function to get a list of all
interfaces currently known to the kernel and operate on that one.